FileSystemTileCache: check for NPE in deleteDirectory

+++ mapsforge-map/src/main/java/org/mapsforge/map/layer/cache/FileSystemTileCache.java
@@ -98,11 +98,13 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
     private static boolean deleteDirectory(File dir) {
         if (dir.isDirectory()) {
             String[] children = dir.list();
-            for (int i = 0; i < children.length; i++) {
-                boolean success = deleteDirectory(new File(dir, children[i]));
-                if (!success) {
-                    return false;
-                }
+            if (children != null) {
+            	for (int i = 0; i < children.length; i++) {
+            		boolean success = deleteDirectory(new File(dir, children[i]));
+            		if (!success) {
+            			return false;
+            		}
+            	}
             }
         }
         // The directory is now empty so delete it
